feat: Update MainView, pass adapter to ViewModel, change constructor if Adapter
Testing in Obsidian itself caused many HTTP-errors in the console and some bugs in the view.
I saw that there is a request-method provided by obsidian to do API requests. Even if fetch etc. worked fine in the unit-test, it causes issues in Obsidian.

Unfortunately this method is missing during unit tests. So I have to pass a parameter for the "do request" method.
In production, the parameter is the method provided by Obsidian.
For unit-tests I can mock this method or do a real implementation like "fetch(request)" etc.
So I'm flexible to do a unit-test or an E2E/integration test.

In addition there were some hiccups with Jest and Webstorm so the tests are commented out. They have to be reworked so they work again with the doRequest-parameter.